- Update to function u2pixel in esru_x.c and esp_draw.c   to explicitly cast return ix and iy parameters to   (long int). This is done in most other places where   fortran calls a C function and expects an integer   returned.
  Note: the fortran source should eventually be
  converted to explicit integer or integer*8 for
  the returned parameter to match the pattern elsewhere.
  
  This commit to test on other platforms.

Testing
- Compile Linux GCC 3.4 X11 (32 bit)
- Interactive test of wireframe view.




Former-commit-id: 9a14cd61b0867bd0b4ac5cf40b68699182015997

+++ src/lib/esru_x.c
@@ -5746,8 +5746,8 @@ void u2pixel_(ux,uy,ix,iy)
   x=(x + x_add) * x_scale;
   y=(y + y_add) * y_scale;
 
-  *ix = x_off + (int) x;
-  *iy = y_off - (int) y;
+  *ix = (long int) x_off + (long int) x;
+  *iy = (long int) y_off - (long int) y;
   return;
 }
